requirements.txt
# ============================================================================
# Gait Analysis FastAPI Server - Optimized Requirements
# Core packages only (removed unnecessary packages for deployment)
# ============================================================================

# ===== FastAPI Web Framework =====
fastapi==0.115.9
uvicorn==0.34.3
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
starlette==0.45.3
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
httptools==0.6.4
watchfiles==1.0.5
websockets==14.2

# ===== LangGraph & LangChain Ecosystem =====
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langchain==0.3.25
langchain-openai==0.3.23
langchain-community==0.3.25
langchain-core==0.3.65
langchain-chroma==0.2.4
langchain-text-splitters==0.3.8
langsmith==0.3.45

# ===== AI & LLM =====
openai==1.86.0
tiktoken==0.9.0

# ===== RAG & Vector Database =====
chromadb==1.0.12
sentence-transformers==4.1.0
pypdf==5.6.0
PyMuPDF==1.26.1
transformers==4.52.4
tokenizers==0.21.1
huggingface-hub==0.33.0
safetensors==0.5.3

# ===== Database & Storage =====
supabase==2.15.3
postgrest==1.0.2
gotrue==2.12.0
realtime==2.4.3
storage3==0.11.3
supafunc==0.9.4
SQLAlchemy==2.0.41

# ===== Deep Learning & Machine Learning (Keras compatibility) =====
tensorflow==2.18.0
tf_keras==2.18.0
keras==3.5.0
scikit-learn==1.6.1
numpy==1.26.4

# ===== Data Processing & Signal Processing =====
pandas==2.2.2
scipy==1.15.3

# ===== Environment Configuration =====
python-dotenv==1.1.0

# ===== HTTP Clients & Utilities =====
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
aiohttp==3.12.13
tqdm==4.67.1

# ===== File Format Support =====
PyYAML==6.0.2

# ===== Required Dependencies =====
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
click==8.2.1
typing_extensions==4.14.0
packaging==24.2
filelock==3.18.0
fsspec==2025.5.1
MarkupSafe==3.0.2
Jinja2==3.1.6
joblib==1.5.1
threadpoolctl==3.6.0
six==1.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
protobuf==4.25.8
grpcio==1.73.0
h5py==3.14.0
absl-py==2.3.0
astunparse==1.6.3
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
libclang==18.1.1
ml-dtypes==0.4.1
opt_einsum==3.4.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
wrapt==1.17.2
regex==2024.11.6
overrides==7.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.25.1
attrs==25.3.0
mmh3==5.1.0
onnxruntime==1.22.0
coloredlogs==15.0.1
humanfriendly==10.0
tenacity==9.1.2
